refactor(engine): move lift_in_subquery into lower
  Move `lift_in_subquery` and `try_lift_relation_path_comparison` from
  `simplify` into `lower::lift_in_subquery`.  Run them as a pre-lowering
  visitor (`LiftInSubquery`) alongside `RewriteVia`, before
  `LowerStatement::visit_stmt_mut`.  Same shape as #817 (via-association)
  rather than the fold-into-`LowerStatement` shape of #823 / #835 / #873.

  The fold pattern would not work here: `ApplyInsertScope::apply_expr`
  walks insert-scope constraint expressions through its own recursion
  and never enters `LowerStatement::visit_expr_mut`.  Without the pre-
  pass, an unlifted relation `IN` subquery reaches `apply_expr` and
  panics on the unsupported shape (177 integration tests).

  Tests migrate from `simplify::tests::lift_in_subquery` to
  `lower::tests::lift_in_subquery`; the helpers are exposed as free
  functions on `&ExprContext` so the unit tests call them without
  constructing a full `LiftInSubquery`.
