+++ crates/toasty/src/engine/lower/tests.rs
@@ -1,2 +1,3 @@
 mod association;
 mod expr_or;
+mod lift_in_subquery;
